+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Usage:
 * Fix #1478: Should detect and warn the user if creating ingress and ingress controller not available
 * Fix #2150: Bump Kubernetes Client to 6.6.0 (fixes issues when trace-logging OpenShift builds)
 * Fix #2162: Bump Kubernetes Client to 6.6.1 (HttpClient with support for PUT + InputStream)
+* Fix #2166: Potential command line injection in SpringBootWatcher
 
 ### 1.12.0 (2023-04-03)
 * Fix #1179: Move storageClass related functionality out of VolumePermissionEnricher to PersistentVolumeClaimStorageClassEnricher